dump-guest-memory.py: convert PhysPageDesc.phys_offset to MPZ (RHEL only)
RH-Author: Laszlo Ersek <lersek@redhat.com>
Message-id: <1395257047-25109-5-git-send-email-lersek@redhat.com>
Patchwork-id: 58148
O-Subject: [RHEL-6.6 qemu-kvm PATCH 4/7] dump-guest-memory.py: convert PhysPageDesc.phys_offset to MPZ (RHEL only)
Bugzilla: 826266
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Amos Kong <akong@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Jeff Nelson <jenelson@redhat.com>

Otherwise "phys_offset", as returned by gdb, trips up the bit-and that
checks the page type encoded in it.

---
 scripts/dump-guest-memory.py | 2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)

+++ scripts/dump-guest-memory.py
@@ -153,7 +153,7 @@ shape and this command should mostly work."""
 
                 for l2 in range (self.L2_SIZE):
                     phys_page_desc = (l2table + l2).dereference()
-                    ram_addr = phys_page_desc["phys_offset"]
+                    ram_addr = long(phys_page_desc["phys_offset"])
 
                     # we only care about RAM
                     if ((ram_addr & ~self.TARGET_PAGE_MASK) != 0):
